@@ -1,4 +1,3 @@
-
 buildscript {
     ext.kotlin_version = '1.3.50'
     repositories {
@@ -13,10 +12,10 @@ buildscript {
 
 allprojects {
     repositories {
-        maven{ url 'https://maven.aliyun.com/repository/google'}
-        maven{ url 'https://maven.aliyun.com/repository/jcenter'}
+        maven { url 'https://maven.aliyun.com/repository/google' }
+        maven { url 'https://maven.aliyun.com/repository/jcenter' }
         maven { url "http://lib.gcssloop.com/repository/gcssloop-central/" }
-        maven {url"https://jitpack.io"}
+        maven { url "https://jitpack.io" }
 
         google()
         jcenter()
@@ -25,8 +24,9 @@ allprojects {
 ext {
     // SDK And Tools
     minSdkVersion = 19
-    targetSdkVersion = 28
-    compileSdkVersion = 28
+    targetSdkVersion = 29
+    compileSdkVersion = 29
+    buildToolsVersion = '29.0.2'
     //Dependencies
     supportLibraryVersion = '28.0.0'
 
