@@ -1,13 +1,13 @@
 package com.example.emoticon.fragment;
 
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v4.widget.SwipeRefreshLayout;
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
+import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -19,8 +19,8 @@ import com.example.common.bean.StatusResult;
 import com.example.common.retrofit.EmoticonProtocol;
 import com.example.common.utils.HttpUtils;
 import com.example.common.utils.ToastUtils;
-import com.example.common.utils.UserManager;
-import com.example.common.widget.Toolbar;
+import com.example.common.manager.UserManager;
+import com.example.common.widget.Topbar;
 import com.example.emoticon.R;
 import com.example.emoticon.activity.EmoticonAddActivity;
 import com.example.emoticon.adapter.EmoticonAdapter;
@@ -35,7 +35,7 @@ public class CreativeFragment extends Fragment implements View.OnClickListener, 
     private SwipeRefreshLayout swipeRefreshLayout;
     int skip = 0;
     EmoticonAdapter adapter;
-    private Toolbar toolbar;
+    private Topbar toolbar;
 
     public static CreativeFragment newInstance(String title) {
         CreativeFragment fragment = new CreativeFragment();
