@@ -3,10 +3,10 @@ package com.example.emoticon.fragment;
 import android.content.Intent;
 import android.graphics.drawable.Drawable;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v4.content.ContextCompat;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.core.content.ContextCompat;
 import android.text.TextUtils;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -18,8 +18,8 @@ import android.widget.TextView;
 import com.bumptech.glide.Glide;
 import com.example.common.bean.User;
 import com.example.common.utils.DrawableTintUtil;
-import com.example.common.utils.UserManager;
-import com.example.common.widget.HeadZoomScrollView;
+import com.example.common.manager.UserManager;
+import com.example.emotion.user.widget.HeadZoomScrollView;
 import com.example.emoticon.R;
 import com.example.emoticon.activity.SettingActivity;
 import com.example.emotion.user.activity.LoginActivity;
