@@ -9,37 +9,28 @@ import android.content.pm.ShortcutManager;
 import android.graphics.drawable.Icon;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import android.view.View;
 import android.widget.EditText;
 
-import com.example.common.RetroClient;
 import com.example.common.app.ResourcesManager;
 import com.example.common.base.BaseActivity;
-import com.example.common.bean.StatusResult;
 import com.example.common.bean.User;
-import com.example.common.utils.HttpUtils;
 import com.example.common.utils.ToastUtils;
-import com.example.common.utils.UserManager;
-import com.example.common.widget.Toolbar;
+import com.example.common.manager.UserManager;
+import com.example.common.widget.Topbar;
 import com.example.emotion.user.R;
-import com.example.emotion.user.retrofit.UserProtocol;
+import com.example.emotion.user.contract.UserContract;
+import com.example.emotion.user.presenter.UserPresenter;
 
-import java.security.KeyFactory;
-import java.security.PublicKey;
-import java.security.spec.X509EncodedKeySpec;
 import java.util.ArrayList;
 import java.util.List;
 
-import javax.crypto.Cipher;
-
-import it.sauronsoftware.base64.Base64;
-import retrofit2.Call;
-
-public class LoginActivity extends BaseActivity implements View.OnClickListener {
-    EditText email, pwd;
-
-    private HttpUtils.RequestFinishCallback<User> mLoginCallback;
+public class LoginActivity extends BaseActivity implements View.OnClickListener, UserContract.LoginView {
+    private EditText email, pwd;
+    private UserContract.Presenter presenter;
+    private ProgressDialog progressDialog;
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
@@ -49,32 +40,14 @@ public class LoginActivity extends BaseActivity implements View.OnClickListener 
     }
 
     private void initViews() {
+        presenter = new UserPresenter(this);
         setTitle("");
         email = findViewById(R.id.email);
         pwd = findViewById(R.id.pwd);
-        Toolbar toolbar = findViewById(R.id.toolbar);
+        Topbar toolbar = findViewById(R.id.toolbar);
         toolbar.setTitle("");
         toolbar.back.setImageResource(R.drawable.close);
         findViewById(R.id.login).setOnClickListener(this);
-    }
-
-
-    //密码公钥加密
-    private String getSessionKey(String pwd) {
-        String sessionKey = "";
-        try {
-            //公钥加密、私钥解密 ---- 加密
-            X509EncodedKeySpec x509EncodedKeySpec2 = new X509EncodedKeySpec(it.sauronsoftware.base64.Base64.decode(RetroClient.PUBLIC_KEY.getBytes()));
-            KeyFactory keyFactory2 = KeyFactory.getInstance("RSA");
-            PublicKey publicKey2 = keyFactory2.generatePublic(x509EncodedKeySpec2);
-            Cipher cipher2 = Cipher.getInstance("RSA");
-            cipher2.init(Cipher.ENCRYPT_MODE, publicKey2);
-            byte[] result2 = cipher2.doFinal(pwd.getBytes());
-            sessionKey = new String(Base64.encode(result2));
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        return sessionKey;
     }
 
     @Override
@@ -85,39 +58,53 @@ public class LoginActivity extends BaseActivity implements View.OnClickListener 
             startActivity(intent);
             finish();
         } else if (id == R.id.login) {
-            if(pwd.getText().toString().isEmpty() || email.getText().toString().isEmpty()){
-                ToastUtils.showToast("不能以空的内容");
+            if (pwd.getText().toString().isEmpty() || email.getText().toString().isEmpty()) {
+                ToastUtils.showToast("不能有空的内容");
                 return;
             }
-            final ProgressDialog progressDialog = new ProgressDialog(this);
+            progressDialog = new ProgressDialog(this);
             progressDialog.setMessage("正在登陆。。");
             progressDialog.show();
-
-            String device = Build.MODEL;
-            String sessionKey = getSessionKey(pwd.getText().toString());
-            UserProtocol userProtocol = RetroClient.getServices(UserProtocol.class);
-            Call<StatusResult<User>> call = userProtocol.login(email.getText().toString(), sessionKey, device);
-            HttpUtils.doRequest(call, new HttpUtils.RequestFinishCallback<User>() {
-                @Override
-                public void getRequest(StatusResult<User> result) {
-                    progressDialog.dismiss();
-                    if (result == null) return;
-                    if (!result.isSuccess()) {
-                        String str = ResourcesManager.getRes().getString(R.string.request_error, result.getMsg());
-                        ToastUtils.showToast(str);
-                        return;
-                    }
-                    if (result.getData() != null) {
-                        UserManager.saveUser(result.getData());
-                        ToastUtils.showToast("登录成功");
-                        shortcuts();
-                        finish();
-                    }
-                }
-            });
+            presenter.login(email.getText().toString(), pwd.getText().toString());
         }
 
     }
+
+    //启动本Activity
+    public static void startActivity(Context context) {
+        Intent intent = new Intent(context, LoginActivity.class);
+        context.startActivity(intent);
+    }
+
+    @Override
+    public void onFinish(@NonNull User user) {
+        if (progressDialog != null) {
+            if (progressDialog.isShowing()) {
+                progressDialog.dismiss();
+            }
+        }
+        UserManager.saveUser(user);
+        ToastUtils.showToast("登录成功");
+        shortcuts();
+        finish();
+    }
+
+    @Override
+    public void onError(@NonNull Exception e) {
+        if (progressDialog != null) {
+            if (progressDialog.isShowing()) {
+                progressDialog.dismiss();
+            }
+        }
+        String str = ResourcesManager.getRes().getString(R.string.request_error, e.getMessage());
+        ToastUtils.showToast(str);
+    }
+
+    @Override
+    public void setPresenter(UserContract.Presenter presenter) {
+        this.presenter = presenter;
+    }
+
 
     //桌面图标长按 ShortCuts
     private void shortcuts() {
@@ -148,9 +135,4 @@ public class LoginActivity extends BaseActivity implements View.OnClickListener 
         }
     }
 
-    //启动本Activity
-    public static void startActivity(Context context) {
-        Intent intent = new Intent(context, LoginActivity.class);
-        context.startActivity(intent);
-    }
 }
