@@ -3,31 +3,27 @@ package com.example.emotion.user.activity;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 
-import com.example.common.RetroClient;
 import com.example.common.app.ResourcesManager;
 import com.example.common.base.BaseActivity;
 import com.example.common.bean.Emoticon;
-import com.example.common.bean.StatusResult;
 import com.example.common.bean.User;
 import com.example.common.fragment.BottomMenuFragmentDialog;
-import com.example.common.utils.HttpUtils;
+import com.example.common.manager.UserManager;
 import com.example.common.utils.ToastUtils;
-import com.example.common.utils.UserManager;
 import com.example.emotion.user.R;
 import com.example.emotion.user.adapter.EmoticonAdapter;
-import com.example.emotion.user.retrofit.UserProtocol;
+import com.example.emotion.user.contract.UserEmoticonsContract;
+import com.example.emotion.user.presenter.UserEmoticonsPresenter;
 
 import java.util.ArrayList;
 import java.util.List;
-
-import retrofit2.Call;
 
 /**
  * 用户表情页面
@@ -36,12 +32,14 @@ import retrofit2.Call;
  * Date: 2019/4/20.
  * PS:
  */
-public class UserEmoticonsActivity extends BaseActivity {
+public class UserEmoticonsActivity extends BaseActivity implements UserEmoticonsContract.EmoticonView {
     User user;
     RecyclerView recyclerView;
     private List<Emoticon> list;
     private EmoticonAdapter adapter;
+    private UserEmoticonsContract.Presenter presenter;
     int skip = 0;
+    private boolean next = false;
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
@@ -58,43 +56,12 @@ public class UserEmoticonsActivity extends BaseActivity {
             return;
         }
         initViews();
-        getData(false);
+        getData();
     }
 
-    private void getData(final boolean next) {
-        if (!next) skip = 0;
-
-        UserProtocol userProtocol = RetroClient.getServices(UserProtocol.class);
-        Call<StatusResult<List<Emoticon>>> emoticonCall = userProtocol.getEmoticonList(user.getId(), 40, skip);
-
-        HttpUtils.doRequest(emoticonCall, new HttpUtils.RequestFinishCallback<List<Emoticon>>() {
-            @Override
-            public void getRequest(StatusResult<List<Emoticon>> result) {
-                if (result == null) return;
-                if (!result.isSuccess()) {
-                    String str = ResourcesManager.getRes().getString(com.example.emotion.user.R.string.request_error, result.getMsg());
-                    ToastUtils.showToast(str);
-                    return;
-                }
-                if (result.getData() != null) {
-                    if (!next) list.clear();
-                    List<Emoticon> data = result.getData();
-                    list.addAll(data);
-//                    adapter.notifyDataSetChanged();
-
-                    adapter.notifyItemRangeInserted(adapter.getItemCount(), data.size());
-                    if (result.getData().size() == 0) {
-                        skip = skip - 40;
-                    } else {
-                        adapter.notifyItemRangeInserted(adapter.getItemCount(), data.size());
-//                        adapter.notifyDataSetChanged();
-                    }
-                }
-            }
-        });
-    }
 
     private void initViews() {
+        presenter = new UserEmoticonsPresenter(this);
         recyclerView = findViewById(R.id.recyclerView);
         list = new ArrayList<>();
         GridLayoutManager gridLayoutManager = new GridLayoutManager(this, 4);
@@ -127,20 +94,24 @@ public class UserEmoticonsActivity extends BaseActivity {
             public void onScrolled(@NonNull RecyclerView recyclerView, int dx, int dy) {
                 super.onScrolled(recyclerView, dx, dy);
                 LinearLayoutManager layoutManager = (LinearLayoutManager) recyclerView.getLayoutManager();
+                if (layoutManager == null) return;
                 int firstCompletelyVisibleItemPosition = layoutManager.findFirstCompletelyVisibleItemPosition();
                 if (firstCompletelyVisibleItemPosition == 0) {
                     //Log.i(TAG, "滑动到顶部");
                 }
-
                 int lastCompletelyVisibleItemPosition = layoutManager.findLastCompletelyVisibleItemPosition();
-                //Log.i(TAG, "lastCompletelyVisibleItemPosition: "+lastCompletelyVisibleItemPosition);
                 if (lastCompletelyVisibleItemPosition == layoutManager.getItemCount() - 1) {
                     skip = skip + 40;
-                    getData(true);
+                    next = true;
+                    getData();
                 }
-                //Log.i(TAG, "滑动到底部");
             }
         });
+    }
+
+    private void getData() {
+        if (!next) skip = 0;
+        presenter.loadData(user.getId(), skip);
     }
 
     public static void startActivity(Context context) {
@@ -148,4 +119,26 @@ public class UserEmoticonsActivity extends BaseActivity {
         context.startActivity(intent);
     }
 
+    @Override
+    public void onError(@NonNull Exception e) {
+        String str = ResourcesManager.getRes().getString(com.example.emotion.user.R.string.request_error, e.getMessage());
+        ToastUtils.showToast(str);
+    }
+
+    @Override
+    public void onFinish(@NonNull List<Emoticon> emoticonList) {
+        if (!next) list.clear();
+        list.addAll(emoticonList);
+        adapter.notifyItemRangeInserted(adapter.getItemCount(), emoticonList.size());
+        if (emoticonList.size() == 0) {
+            skip = skip - 40;
+        } else {
+            adapter.notifyItemRangeInserted(adapter.getItemCount(), emoticonList.size());
+        }
+    }
+
+    @Override
+    public void setPresenter(UserEmoticonsContract.Presenter presenter) {
+        this.presenter = presenter;
+    }
 }
