@@ -7,11 +7,6 @@ import android.content.DialogInterface;
 import android.content.Intent;
 import android.graphics.Color;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.text.InputType;
 import android.text.TextUtils;
 import android.view.LayoutInflater;
@@ -19,6 +14,12 @@ import android.view.View;
 import android.widget.EditText;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AlertDialog;
+import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.cy.cyflowlayoutlibrary.FlowLayout;
 import com.cy.cyflowlayoutlibrary.FlowLayoutAdapter;
@@ -28,14 +29,14 @@ import com.example.common.app.ResourcesManager;
 import com.example.common.base.BaseActivity;
 import com.example.common.bean.EmoticonType;
 import com.example.common.bean.StatusResult;
+import com.example.common.manager.UserManager;
 import com.example.common.particlesmaster.ParticleSmasher;
 import com.example.common.particlesmaster.SmashAnimator;
 import com.example.common.retrofit.EmoticonProtocol;
 import com.example.common.retrofit.EmoticonTypeProtocol;
 import com.example.common.utils.HttpUtils;
 import com.example.common.utils.ToastUtils;
-import com.example.common.utils.UserManager;
-import com.example.common.widget.Toolbar;
+import com.example.common.widget.Topbar;
 import com.example.emoticon.R;
 import com.example.emoticon.adapter.EmoticonAddImageAdapter;
 import com.example.emoticon.adapter.LabelAdapter;
@@ -66,7 +67,7 @@ public class EmoticonAddActivity extends BaseActivity implements View.OnClickLis
     List<String> tagList = new ArrayList<>();
     List<String> imgList = new ArrayList<>();
     private FlowLayoutAdapter<String> flowLayoutAdapter;
-    Toolbar toolbar;
+    Topbar toolbar;
     ImageView imageSelect;
     int typeId = 0;
     String type_title;
