@@ -1,11 +1,12 @@
 package com.example.emoticon.activity;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
 import android.view.View;
 
+import androidx.annotation.Nullable;
+
 import com.example.common.base.BaseActivity;
-import com.example.common.utils.UserManager;
+import com.example.common.manager.UserManager;
 import com.example.emoticon.R;
 
 public class SettingActivity extends BaseActivity implements View.OnClickListener {
