@@ -1,17 +1,18 @@
 package com.example.emoticon.adapter;
 
 import android.graphics.Bitmap;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.view.HapticFeedbackConstants;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.bumptech.glide.Glide;
 import com.bumptech.glide.request.target.SimpleTarget;
@@ -20,6 +21,7 @@ import com.example.common.bean.Emoticon;
 import com.example.common.fragment.BottomMenuFragmentDialog;
 import com.example.common.utils.ImageUtils;
 import com.example.emoticon.R;
+import com.example.qrcode.BitmapUtils;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -113,8 +115,8 @@ public class EmoticonAdapter extends RecyclerView.Adapter<EmoticonAdapter.ViewHo
         }
     }
 
-    public Bitmap getBitMap(int position) {
-        return bitmaps.get(position);
+    public String getBitMapCachePath(int position) {
+        return BitmapUtils.saveBitmapToCacheDir(bitmaps.get(position));
     }
 
     public interface OnItemClickListener {
