@@ -2,14 +2,15 @@ package com.example.emoticon.fragment;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.common.RetroClient;
 import com.example.common.app.ResourcesManager;
@@ -109,15 +110,12 @@ public class DetailsFragment extends Fragment {
 //            }
 //        });
 
-        adapter.setOnItemClickListener(new EmoticonAdapter.OnItemClickListener() {
-            @Override
-            public void onItemClick(View view, int position) {
+        adapter.setOnItemClickListener((v,position)-> {
                 Intent intent = new Intent(getActivity(), EditActivity.class);
                 String image = list.get(position).getImgUrl() + ImageUtils.gifToJpg;
                 intent.putExtra("picture", image);
-                intent.putExtra("bitmap", adapter.getBitMap(position));
+                intent.putExtra("bitmap", adapter.getBitMapCachePath(position));
                 startActivity(intent);
-            }
         });
         return view;
     }
