@@ -8,12 +8,6 @@ import android.net.Uri;
 import android.os.Bundle;
 import android.os.Environment;
 import android.provider.MediaStore;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.design.widget.TabLayout;
-import android.support.v4.app.Fragment;
-import android.support.v4.view.ViewPager;
-import android.support.v7.widget.RecyclerView;
 import android.view.Gravity;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -22,16 +16,23 @@ import android.view.WindowManager;
 import android.widget.PopupWindow;
 import android.widget.TextView;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.viewpager.widget.ViewPager;
+
 import com.example.common.RetroClient;
 import com.example.common.bean.Emoticon;
 import com.example.common.bean.EmoticonType;
 import com.example.common.retrofit.EmoticonTypeProtocol;
 import com.example.common.utils.ToastUtils;
-import com.example.common.widget.Toolbar;
+import com.example.common.widget.Topbar;
 import com.example.emoticon.R;
 import com.example.emoticon.adapter.MainPageAdapter;
 import com.example.emoticon.editmodule.activity.EditActivity;
 import com.example.emoticon.utils.ScreenUtils;
+import com.google.android.material.tabs.TabLayout;
 
 import java.io.ByteArrayOutputStream;
 import java.util.ArrayList;
@@ -74,7 +75,7 @@ public class PaintFragment extends Fragment implements View.OnClickListener {
         mViewPager.setAdapter(adapter);
         mTabLayout.setupWithViewPager(mViewPager);
 
-        Toolbar toolbar = view.findViewById(R.id.toolbar);
+        Topbar toolbar = view.findViewById(R.id.toolbar);
         toolbar.setRightButtonOneShow(true);
         toolbar.right1.setImageResource(R.drawable.cam);
         toolbar.right1.setOnClickListener(new View.OnClickListener() {
