@@ -10,11 +10,10 @@ import retrofit2.converter.gson.GsonConverterFactory;
 
 public class RetroClient {
     //服务器地址
-    public static final String PUBLIC_KEY = "MIGfMA0GCSqGSIb3DQEBAQUAA4GNADCBiQKBgQCkDtr9lR2GLoGeJWLmXHbf1HSEvOzH+Wgz7+Dc9nTEKNXnSq1g8c5UmG4YbXwUtuFwJMgR06GWkMCX3yqXBeXKM/wibQLRr/5UGVNJgUo8jpajwScpuzqb1QLohluyhRd0TKp7JJDqpOIWsIKrc6c4qx/MeDZYJg2cXwH829T7HQIDAQAB";
     private static final OkHttpClient client = new OkHttpClient.Builder().
             connectTimeout(60, TimeUnit.SECONDS).
             readTimeout(60, TimeUnit.SECONDS).
-            callTimeout(60,TimeUnit.SECONDS).
+            callTimeout(60, TimeUnit.SECONDS).
             writeTimeout(60, TimeUnit.SECONDS).build();
 
     //获取Retrofit
@@ -26,7 +25,7 @@ public class RetroClient {
                 .build();
     }
 
-    public static <T> T getServices(Class<T> cls){
+    public static <T> T getServices(Class<T> cls) {
         Retrofit retrofit = RetroClient.getRetroClient();
         return retrofit.create(cls);
     }
